CSCFAIRMETA-806: Use Finnish title in Qvain when available * When possible, show title in the datasets view using current locale * Rename language.js -> locale.js to match the exported component
