+++ etsin_finder/frontend/__tests__/cumulativeState.test.jsx
@@ -10,7 +10,7 @@ import etsinTheme from '../js/styles/theme'
 import { CUMULATIVE_STATE } from '../js/utils/constants'
 import Env from '../js/stores/domain/env'
 import QvainStoreClass from '../js/stores/view/qvain'
-import Locale from '../js/stores/view/language'
+import Locale from '../js/stores/view/locale'
 import CumulativeState, {
   CumulativeStateButton,
 } from '../js/components/qvain/fields/files/cumulativeStateV2'
