+++ etsin_finder/frontend/__tests__/qvain.datepicker.test.jsx
@@ -6,7 +6,7 @@ import { ThemeProvider } from 'styled-components'
 import '../locale/translations'
 import etsinTheme from '../js/styles/theme'
 import DatePicker, { StyledCustomDatePicker } from '../js/components/qvain/general/input/datepicker'
-import LocaleStore from '../js/stores/view/language'
+import LocaleStore from '../js/stores/view/locale'
 
 global.Promise = require('bluebird')
