+++ etsin_finder/frontend/__tests__/qvain.files.legacy.jsx
@@ -17,7 +17,7 @@ import { DeleteButton } from '../js/components/qvain/general/buttons'
 import Env from '../js/stores/domain/env'
 import QvainStoreClass from '../js/stores/view/qvain'
 import { Directory } from '../js/stores/view/qvain/qvain.filesv1'
-import LocaleStore from '../js/stores/view/language'
+import LocaleStore from '../js/stores/view/locale'
 import { DATA_CATALOG_IDENTIFIER } from '../js/utils/constants'
 import { useStores } from '../js/stores/stores'
