+++ etsin_finder/frontend/__tests__/qvain.metadata.store.test.js
@@ -1,7 +1,7 @@
 import axios from 'axios'
 import Env from '../js/stores/domain/env'
 import QvainStoreClass from '../js/stores/view/qvain'
-import LocaleStore from '../js/stores/view/language'
+import LocaleStore from '../js/stores/view/locale'
 import { Directory } from '../js/stores/view/qvain/qvain.filesv1'
 
 jest.mock('axios')
