+++ etsin_finder/frontend/__tests__/qvainPas.legacy.test.jsx
@@ -19,7 +19,7 @@ import FileForm from '../js/components/qvain/fields/files/legacy/fileForm'
 import IDAFilePicker from '../js/components/qvain/fields/files/legacy/idaFilePicker'
 import QvainStoreClass from '../js/stores/view/qvain'
 import { Directory, File } from '../js/stores/view/qvain/qvain.filesv1'
-import LocaleStore from '../js/stores/view/language'
+import LocaleStore from '../js/stores/view/locale'
 import EnvStore from '../js/stores/domain/env'
 import { ACCESS_TYPE_URL, DATA_CATALOG_IDENTIFIER } from '../js/utils/constants'
 import { StoresProvider, useStores } from '../js/stores/stores'
