+++ etsin_finder/frontend/__tests__/qvainPas.test.jsx
@@ -19,7 +19,7 @@ import FileForm from '../js/components/qvain/fields/files/ida/forms/fileForm'
 import DirectoryForm from '../js/components/qvain/fields/files/ida/forms/directoryForm'
 import { File, Directory, Project } from '../js/stores/view/common.files.items'
 import QvainStoreClass from '../js/stores/view/qvain'
-import LocaleStore from '../js/stores/view/language'
+import LocaleStore from '../js/stores/view/locale'
 import EnvStore from '../js/stores/domain/env'
 import { ACCESS_TYPE_URL, DATA_CATALOG_IDENTIFIER } from '../js/utils/constants'
 import { metaxResponses } from './__testdata__/qvainPas.data'
