+++ etsin_finder/frontend/js/stores/view/locale.js
