+++ etsin_finder/frontend/js/utils/checkDataLang.jsx
@@ -13,7 +13,7 @@
 // checks if there is a translation available for specific language.
 // if no language is specified, uses current language.
 
-import Locale from '../stores/view/language'
+import Locale from '../stores/view/locale'
 
 const checkDataLang = (object, lang) => {
   let language = lang
